reworked a lot of logic to do with waypoint placement rules + building pathways and logic to determine if dog shoould use them or not.
added extentions to player controller to get dog fall speed and touching ground check working when not playing as dog + a copy robot jump delay.

dog now properly follows player with jumps.
waypoint system still needs work with when deciding to use it or not + when pathing needs to be checked and updates.