fixed more bugs and reworked logic that dictates if dog should use waypoints or not. no obvious bugs anymore with it getting stuck and will always find its way back to player.
also fixed dog jumping when it should both while following and waypoint following with grounded checks. small issue with dog failing jumps due to hitting head on the bottom of platforms especially with player standing on it.

issues with dynamic waypoint placing. failing jumps can make a waypoint be placed when it shouldnt.possible fixed listed in WaypointManager script.

updated scene manager to set the most recently loaded scene as the active scene